+++ packages/drivekit_core/flutter_drivekit_core_ios/lib/src/core_api.g.dart
@@ -15,7 +15,8 @@ PlatformException _createConnectionError(String channelName) {
   );
 }
 
-List<Object?> wrapResponse({Object? result, PlatformException? error, bool empty = false}) {
+List<Object?> wrapResponse(
+    {Object? result, PlatformException? error, bool empty = false}) {
   if (empty) {
     return <Object?>[];
   }
@@ -76,6 +77,7 @@ enum PigeonDeviceConfigurationEvent {
 enum PigeonSynchronizationType {
   /// synchronize by calling the DriveQuant servers
   defaultSync,
+
   /// retrieve already synchronized items in the local database
   cache,
 }
@@ -84,8 +86,10 @@ enum PigeonSynchronizationType {
 enum PigeonUserInfoSyncStatus {
   /// Synchronization has been successfully performed
   success,
+
   /// SynchronizationType has been set to cache.
   cacheDataOnly,
+
   /// Synchronization has failed,
   /// only user info previously synchronized is returned
   failedToSyncUserInfoCacheOnly,
@@ -153,7 +157,6 @@ class PigeonGetUserInfoResponse {
   }
 }
 
-
 class _PigeonCodec extends StandardMessageCodec {
   const _PigeonCodec();
   @override
@@ -161,28 +164,28 @@ class _PigeonCodec extends StandardMessageCodec {
     if (value is PigeonUserInfo) {
       buffer.putUint8(129);
       writeValue(buffer, value.encode());
-    } else     if (value is PigeonGetUserInfoResponse) {
+    } else if (value is PigeonGetUserInfoResponse) {
       buffer.putUint8(130);
       writeValue(buffer, value.encode());
-    } else     if (value is PigeonDeleteAccountStatus) {
+    } else if (value is PigeonDeleteAccountStatus) {
       buffer.putUint8(131);
       writeValue(buffer, value.index);
-    } else     if (value is PigeonRequestError) {
+    } else if (value is PigeonRequestError) {
       buffer.putUint8(132);
       writeValue(buffer, value.index);
-    } else     if (value is PigeonUpdateUserIdStatus) {
+    } else if (value is PigeonUpdateUserIdStatus) {
       buffer.putUint8(133);
       writeValue(buffer, value.index);
-    } else     if (value is PigeonBackgroundFetchStatus) {
+    } else if (value is PigeonBackgroundFetchStatus) {
       buffer.putUint8(134);
       writeValue(buffer, value.index);
-    } else     if (value is PigeonDeviceConfigurationEvent) {
+    } else if (value is PigeonDeviceConfigurationEvent) {
       buffer.putUint8(135);
       writeValue(buffer, value.index);
-    } else     if (value is PigeonSynchronizationType) {
+    } else if (value is PigeonSynchronizationType) {
       buffer.putUint8(136);
       writeValue(buffer, value.index);
-    } else     if (value is PigeonUserInfoSyncStatus) {
+    } else if (value is PigeonUserInfoSyncStatus) {
       buffer.putUint8(137);
       writeValue(buffer, value.index);
     } else {
@@ -193,29 +196,31 @@ class _PigeonCodec extends StandardMessageCodec {
   @override
   Object? readValueOfType(int type, ReadBuffer buffer) {
     switch (type) {
-      case 129: 
+      case 129:
         return PigeonUserInfo.decode(readValue(buffer)!);
-      case 130: 
+      case 130:
         return PigeonGetUserInfoResponse.decode(readValue(buffer)!);
-      case 131: 
+      case 131:
         final int? value = readValue(buffer) as int?;
         return value == null ? null : PigeonDeleteAccountStatus.values[value];
-      case 132: 
+      case 132:
         final int? value = readValue(buffer) as int?;
         return value == null ? null : PigeonRequestError.values[value];
-      case 133: 
+      case 133:
         final int? value = readValue(buffer) as int?;
         return value == null ? null : PigeonUpdateUserIdStatus.values[value];
-      case 134: 
+      case 134:
         final int? value = readValue(buffer) as int?;
         return value == null ? null : PigeonBackgroundFetchStatus.values[value];
-      case 135: 
+      case 135:
         final int? value = readValue(buffer) as int?;
-        return value == null ? null : PigeonDeviceConfigurationEvent.values[value];
-      case 136: 
+        return value == null
+            ? null
+            : PigeonDeviceConfigurationEvent.values[value];
+      case 136:
         final int? value = readValue(buffer) as int?;
         return value == null ? null : PigeonSynchronizationType.values[value];
-      case 137: 
+      case 137:
         final int? value = readValue(buffer) as int?;
         return value == null ? null : PigeonUserInfoSyncStatus.values[value];
       default:
@@ -228,9 +233,11 @@ class IOSCoreApi {
   /// Constructor for [IOSCoreApi].  The [binaryMessenger] named argument is
   /// available for dependency injection.  If it is left null, the default
   /// BinaryMessenger will be used which routes to the host platform.
-  IOSCoreApi({BinaryMessenger? binaryMessenger, String messageChannelSuffix = ''})
+  IOSCoreApi(
+      {BinaryMessenger? binaryMessenger, String messageChannelSuffix = ''})
       : __pigeon_binaryMessenger = binaryMessenger,
-        __pigeon_messageChannelSuffix = messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
+        __pigeon_messageChannelSuffix =
+            messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
   final BinaryMessenger? __pigeon_binaryMessenger;
 
   static const MessageCodec<Object?> pigeonChannelCodec = _PigeonCodec();
@@ -238,8 +245,10 @@ class IOSCoreApi {
   final String __pigeon_messageChannelSuffix;
 
   Future<void> setApiKey(String key) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.setApiKey$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.setApiKey$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -260,8 +269,10 @@ class IOSCoreApi {
   }
 
   Future<void> setUserId(String userId) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.setUserId$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.setUserId$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -282,8 +293,10 @@ class IOSCoreApi {
   }
 
   Future<String?> getUserId() async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getUserId$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getUserId$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -304,8 +317,10 @@ class IOSCoreApi {
   }
 
   Future<void> updateUserId(String userId) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.updateUserId$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.updateUserId$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -326,8 +341,10 @@ class IOSCoreApi {
   }
 
   Future<String?> getInstallationId() async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getInstallationId$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getInstallationId$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -347,15 +364,19 @@ class IOSCoreApi {
     }
   }
 
-  Future<PigeonGetUserInfoResponse> getUserInfo({PigeonSynchronizationType synchronizationType = PigeonSynchronizationType.defaultSync}) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getUserInfo$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+  Future<PigeonGetUserInfoResponse> getUserInfo(
+      {PigeonSynchronizationType synchronizationType =
+          PigeonSynchronizationType.defaultSync}) async {
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getUserInfo$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
     );
-    final List<Object?>? __pigeon_replyList =
-        await __pigeon_channel.send(<Object?>[synchronizationType]) as List<Object?>?;
+    final List<Object?>? __pigeon_replyList = await __pigeon_channel
+        .send(<Object?>[synchronizationType]) as List<Object?>?;
     if (__pigeon_replyList == null) {
       throw _createConnectionError(__pigeon_channelName);
     } else if (__pigeon_replyList.length > 1) {
@@ -375,8 +396,10 @@ class IOSCoreApi {
   }
 
   Future<bool> updateUserInfo(PigeonUserInfo userInfo) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.updateUserInfo$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.updateUserInfo$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -402,8 +425,10 @@ class IOSCoreApi {
   }
 
   Future<void> reset() async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.reset$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.reset$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -424,8 +449,10 @@ class IOSCoreApi {
   }
 
   Future<bool> isTokenValid() async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.isTokenValid$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.isTokenValid$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -451,14 +478,16 @@ class IOSCoreApi {
   }
 
   Future<void> deleteAccount({bool instantDeletion = false}) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.deleteAccount$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.deleteAccount$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
     );
-    final List<Object?>? __pigeon_replyList =
-        await __pigeon_channel.send(<Object?>[instantDeletion]) as List<Object?>?;
+    final List<Object?>? __pigeon_replyList = await __pigeon_channel
+        .send(<Object?>[instantDeletion]) as List<Object?>?;
     if (__pigeon_replyList == null) {
       throw _createConnectionError(__pigeon_channelName);
     } else if (__pigeon_replyList.length > 1) {
@@ -473,8 +502,10 @@ class IOSCoreApi {
   }
 
   Future<String?> getApiKey() async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getApiKey$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getApiKey$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -495,8 +526,10 @@ class IOSCoreApi {
   }
 
   Future<void> enableLogging({bool showInConsole = true}) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.enableLogging$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.enableLogging$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -517,8 +550,10 @@ class IOSCoreApi {
   }
 
   Future<void> disableLogging({bool showInConsole = true}) async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.disableLogging$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.disableLogging$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -539,8 +574,10 @@ class IOSCoreApi {
   }
 
   Future<String?> getLogUriFile() async {
-    final String __pigeon_channelName = 'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getLogUriFile$__pigeon_messageChannelSuffix';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+    final String __pigeon_channelName =
+        'dev.flutter.pigeon.pigeon_core_package.IOSCoreApi.getLogUriFile$__pigeon_messageChannelSuffix';
+    final BasicMessageChannel<Object?> __pigeon_channel =
+        BasicMessageChannel<Object?>(
       __pigeon_channelName,
       pigeonChannelCodec,
       binaryMessenger: __pigeon_binaryMessenger,
@@ -570,7 +607,8 @@ abstract class FlutterCoreApi {
 
   void driveKitDidReceiveAuthenticationError(PigeonRequestError error);
 
-  void userIdUpdateStatusChanged(PigeonUpdateUserIdStatus status, String? userId);
+  void userIdUpdateStatusChanged(
+      PigeonUpdateUserIdStatus status, String? userId);
 
   void driveKitAccountDeletionCompleted(PigeonDeleteAccountStatus status);
 
@@ -578,11 +616,18 @@ abstract class FlutterCoreApi {
 
   void onDeviceConfigurationChanged(PigeonDeviceConfigurationEvent event);
 
-  static void setUp(FlutterCoreApi? api, {BinaryMessenger? binaryMessenger, String messageChannelSuffix = '',}) {
-    messageChannelSuffix = messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
+  static void setUp(
+    FlutterCoreApi? api, {
+    BinaryMessenger? binaryMessenger,
+    String messageChannelSuffix = '',
+  }) {
+    messageChannelSuffix =
+        messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
     {
-      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidConnect$messageChannelSuffix', pigeonChannelCodec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidConnect$messageChannelSuffix',
+          pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
         __pigeon_channel.setMessageHandler(null);
@@ -593,15 +638,18 @@ abstract class FlutterCoreApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          }          catch (e) {
-            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
           }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidDisconnect$messageChannelSuffix', pigeonChannelCodec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidDisconnect$messageChannelSuffix',
+          pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
         __pigeon_channel.setMessageHandler(null);
@@ -612,24 +660,28 @@ abstract class FlutterCoreApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          }          catch (e) {
-            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
           }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidReceiveAuthenticationError$messageChannelSuffix', pigeonChannelCodec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidReceiveAuthenticationError$messageChannelSuffix',
+          pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
         __pigeon_channel.setMessageHandler(null);
       } else {
         __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
-          'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidReceiveAuthenticationError was null.');
+              'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidReceiveAuthenticationError was null.');
           final List<Object?> args = (message as List<Object?>?)!;
-          final PigeonRequestError? arg_error = (args[0] as PigeonRequestError?);
+          final PigeonRequestError? arg_error =
+              (args[0] as PigeonRequestError?);
           assert(arg_error != null,
               'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitDidReceiveAuthenticationError was null, expected non-null PigeonRequestError.');
           try {
@@ -637,24 +689,28 @@ abstract class FlutterCoreApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          }          catch (e) {
-            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
           }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.userIdUpdateStatusChanged$messageChannelSuffix', pigeonChannelCodec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.userIdUpdateStatusChanged$messageChannelSuffix',
+          pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
         __pigeon_channel.setMessageHandler(null);
       } else {
         __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
-          'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.userIdUpdateStatusChanged was null.');
+              'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.userIdUpdateStatusChanged was null.');
           final List<Object?> args = (message as List<Object?>?)!;
-          final PigeonUpdateUserIdStatus? arg_status = (args[0] as PigeonUpdateUserIdStatus?);
+          final PigeonUpdateUserIdStatus? arg_status =
+              (args[0] as PigeonUpdateUserIdStatus?);
           assert(arg_status != null,
               'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.userIdUpdateStatusChanged was null, expected non-null PigeonUpdateUserIdStatus.');
           final String? arg_userId = (args[1] as String?);
@@ -663,24 +719,28 @@ abstract class FlutterCoreApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          }          catch (e) {
-            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
           }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitAccountDeletionCompleted$messageChannelSuffix', pigeonChannelCodec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitAccountDeletionCompleted$messageChannelSuffix',
+          pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
         __pigeon_channel.setMessageHandler(null);
       } else {
         __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
-          'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitAccountDeletionCompleted was null.');
+              'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitAccountDeletionCompleted was null.');
           final List<Object?> args = (message as List<Object?>?)!;
-          final PigeonDeleteAccountStatus? arg_status = (args[0] as PigeonDeleteAccountStatus?);
+          final PigeonDeleteAccountStatus? arg_status =
+              (args[0] as PigeonDeleteAccountStatus?);
           assert(arg_status != null,
               'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitAccountDeletionCompleted was null, expected non-null PigeonDeleteAccountStatus.');
           try {
@@ -688,24 +748,28 @@ abstract class FlutterCoreApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          }          catch (e) {
-            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
           }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitBackgroundFetchStatusChanged$messageChannelSuffix', pigeonChannelCodec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitBackgroundFetchStatusChanged$messageChannelSuffix',
+          pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
         __pigeon_channel.setMessageHandler(null);
       } else {
         __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
-          'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitBackgroundFetchStatusChanged was null.');
+              'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitBackgroundFetchStatusChanged was null.');
           final List<Object?> args = (message as List<Object?>?)!;
-          final PigeonBackgroundFetchStatus? arg_status = (args[0] as PigeonBackgroundFetchStatus?);
+          final PigeonBackgroundFetchStatus? arg_status =
+              (args[0] as PigeonBackgroundFetchStatus?);
           assert(arg_status != null,
               'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.driveKitBackgroundFetchStatusChanged was null, expected non-null PigeonBackgroundFetchStatus.');
           try {
@@ -713,24 +777,28 @@ abstract class FlutterCoreApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          }          catch (e) {
-            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
           }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.onDeviceConfigurationChanged$messageChannelSuffix', pigeonChannelCodec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.onDeviceConfigurationChanged$messageChannelSuffix',
+          pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
         __pigeon_channel.setMessageHandler(null);
       } else {
         __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
-          'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.onDeviceConfigurationChanged was null.');
+              'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.onDeviceConfigurationChanged was null.');
           final List<Object?> args = (message as List<Object?>?)!;
-          final PigeonDeviceConfigurationEvent? arg_event = (args[0] as PigeonDeviceConfigurationEvent?);
+          final PigeonDeviceConfigurationEvent? arg_event =
+              (args[0] as PigeonDeviceConfigurationEvent?);
           assert(arg_event != null,
               'Argument for dev.flutter.pigeon.pigeon_core_package.FlutterCoreApi.onDeviceConfigurationChanged was null, expected non-null PigeonDeviceConfigurationEvent.');
           try {
@@ -738,8 +806,9 @@ abstract class FlutterCoreApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          }          catch (e) {
-            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
           }
         });
       }